requirements.txt
# AI和语音处理相关依赖
omegaconf~=2.3.0
torch~=2.1.0
tqdm
einops
vector_quantize_pytorch
transformers~=4.41.1
vocos
IPython
whisper~=1.1.10
openai~=1.3.0
edge-tts~=6.1.9
python-dotenv~=1.0.0

# 音频处理相关依赖
pygame~=2.5.2
scipy~=1.11.3

# MQTT通信相关
siot

# 其他工具依赖
numpy~=1.24.3